SymetricCryptography/requirements.txt
cryptography==37.0.2
matplotlib==3.5.2
numpy==1.22.4
Pillow==9.1.1
scikit_image==0.19.2
